Adding Grokking Algorithms to the list of books
Hi,
I am the author of Grokking Algorithms. I have been writing this book for almost three years and I'm quite proud of it. I think it belongs on this list! So far readers have found it to be really easy to read:

"It makes a usually dry and intimidating topic easily digestible and understandable. It is a must read!"

"This is the book I’ve been waiting for to share with both my kids who are interested in programming AND my adult learners attending the various introductory programming meetups I run."

Other reader comments: manning-sandbox.com/forum.jspa?forumID=920