Add Toph to list of online judges
Toph is a relatively new (a little over an year old) sport programming platform. In this time, Toph has hosted quite a number of programming contests. It's arena is real-time, making the contestants' experience as pleasant as possible. Toph also has a growing problem archive that allows problem solvers to practice and improve their skills. Beyond these, Toph's other functionalities (ability to form teams and participant in contests as team) builds on that experience. Hosting a contest on Toph is also very enjoyable with its rich contest dashboard for administrators and contest moderators, integration with external tools like Slack, and so on.